functions/zip-to-parquet-python/requirements.txt
# Function dependencies, for example:
# google-cloud-storage
# fiona

google-cloud-storage==2.16.0
Fiona>=1.9